requirements.txt
crewai==0.51.0
langchain==0.3.0
langchain-community
langchain-openai
chromadb
fastapi
uvicorn[standard]
celery[redis]
redis
python-dotenv
duckduckgo-search
streamlit
pydantic
